requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.15.2
amqp==5.3.1
analytics-python==1.4.post1
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
attrs==25.3.0
Authlib==1.6.1
azure-core==1.33.0
azure-storage-blob==12.25.1
backoff==1.10.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
Brotli @ file:///croot/brotli-split_1736182456865/work
browser-use==0.7.7
bs4==0.0.2
bubus==1.5.6
cachetools==5.5.2
cdp-use==1.4.0
celery==5.5.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docstring_parser @ file:///home/conda/feedstock_root/build_artifacts/docstring_parser_1753195641777/work
dotenv==0.9.9
email-reply-parser==0.5.12
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.115.12
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
google-adk==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core @ file:///home/conda/feedstock_root/build_artifacts/google-api-core-split_1749796131763/work
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform @ file:///home/conda/feedstock_root/build_artifacts/google-cloud-aiplatform_1758097790999/work
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery @ file:///home/conda/feedstock_root/build_artifacts/google-cloud-bigquery-split_1757493843908/work
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.36.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio_client @ file:///home/conda/feedstock_root/build_artifacts/gradio-client_1757549658293/work
graphviz==0.21
greenlet==3.2.1
groovy @ file:///home/conda/feedstock_root/build_artifacts/groovy-python_1753106748838/work
groq==0.31.1
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.1.9
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html2text==2025.4.15
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1758097109572/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
invoke==2.2.0
isodate==0.7.2
itsdangerous @ file:///croot/itsdangerous_1716533337253/work
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
kiwisolver==1.4.9
kombu==5.5.3
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.14
linkify-it-py==2.0.3
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.12.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
monotonic==1.6
motor==3.7.1
multidict==6.6.3
mypy_extensions==1.1.0
netifaces==0.10.6
numpy==2.2.5
oauthlib==3.2.2
ollama==0.5.3
openai==1.107.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
paramiko==4.0.0
parse==1.20.2
pillow==11.2.1
platformdirs==4.5.0
playwright==1.51.0
pluggy==1.6.0
portalocker==2.10.1
posthog==6.7.4
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg-binary==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.12.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.3
pypdf==5.7.0
pyppeteer==0.0.25
pyquery==2.0.1
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
r2r==3.5.15
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.13.0
safehttpx==0.1.6
screeninfo==0.8.1
selenium==4.31.0
semantic-version @ file:///home/conda/feedstock_root/build_artifacts/semantic_version_1653579368137/work
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==3.0.2
starlette==0.46.2
tavily-python==0.7.10
telnyx==2.1.6
tenacity==8.5.0
termcolor==3.0.1
textual==6.3.0
tiktoken==0.8.0
toml==0.10.2
tomlkit==0.13.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.7.0
typer==0.17.4
types-aiofiles==24.1.0.20250326
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
utils==1.0.2
uuid7==0.1.0
uvicorn==0.34.2
vine==5.1.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
